Use json serialization of enabled_methods
See issue #56

If the methods listed in enabled_methods are unicode (for example the
unicode_literals future is enabled in settings.py where SWAGGER_SETTINGS
is defined) then mark_safe will output u'get', which is not valid
javascript.

This commit uses json to output the enabled_methods, which will always
be javascript parseable.